Return size_t from string length function
Both `strlen` and `strlen_P` return `size_t`.
`size_t` is the proper type for indicating the size of a data structure (in bytes or characters).

+++ Source/Ardutosh/StringWrapper.cpp
@@ -3,7 +3,7 @@
 #include "Defines.h"
 #include <string.h>
 
-int xString::CalculateLength() const 
+size_t xString::CalculateLength() const 
 {
 	if (!data)
 		return 0;
